added links to my public github
This is the most important part of the website. I have links to my
Github, which will allow employers to look at the code I have written
for various courses. This abides by the cs.columbia.edu academic honesty
policies. For certain courses, I can only use a secret github gist,
since those are not publicly searchable.

This is still a work in progress.