This webpage will hold videos from my playing days
This code was by far the hardest thing to write so far. It included
writing css classes, and javascript events that would allow me to create
a gallery of 8 hitting video highlights.

There is a main video that is featured, and then to the right there is a
list that includes 7 other highlight videos in a menu.

This code was hard, but I think it came out pretty well. It makes sense
for me to have all of these highlights on another page, I didn't want it
to clutter the 'about me' page.