added new css for video styler
I have adopted the grid layout for all of my webpages, and in order
to do that I needed more css code. I split the page into columns and
rows and then append content as needed.

These containers (used in the grid) also resize pretty well on mobile.

Not perfect, but getting there.